\name{givedim}
\alias{givedim}
\title{Give all the elements in a dimension selection}
\description{Split numbers (separated by : and ,) in a string}
\usage{givedim(v)}
\arguments{
  \item{v}{character string with numbers separated by : and ,}
}
\details{Used to interpret the dimension indicated in a input box}
\value{Numeric vector of integers}
\author{Gianmarco Polotti}
\examples{givedim('1:3,8,7')}
\keyword{ Assignment }
