\name{givedim}
\alias{givedim}
\title{Give full element in dimension selection}
\description{Split e detail the numbers in a short sequence made by : and ,}
\usage{givedim(v)}
\arguments{
  \item{v}{character string with numbers separated by : and ,}
}
\details{Used to interpret the dimension indicated in a input box}
\value{numeric vector of integers}
\author{Gianmarco Polotti}
\examples{givedim('1:3,8,7')}
\keyword{ input box}
