\name{menucollection-package}
\alias{menucollection-package}
\alias{menucollection}
\docType{package}
\title{Collection of ready to use graphical menus}
\description{Graphical menus developed with the GTK+ library configurable by the user}
\details{
\tabular{ll}{
Package: \tab menucollection\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-19\cr
License: \tab GPL-2\cr
}
}
\author{Gianmarco Polotti}
\keyword{ package }
